requirements.txt
Django==3.0
djangorestframework==3.11.0
django-filter==2.2.0
gunicorn==20.0.4
mysqlclient==1.4.6
django-debug-toolbar==2.1
django-cors-headers==3.2.1